+++ Project2/Scene/TitleScene.h
@@ -15,7 +15,7 @@ public:
 	virtual unipueBase Update(unipueBase own) override final;
 private:
 	std::vector<sharedButton> button_;
-	size_t cursorNum_;
+	int cursorNum_;
 	static bool close_;
 	PlayEreaVec playErea_;
 };
